chore(custom-metrics): rebase and update branch
- rebase and update branch
- move test to integration tests directory
- update dependencies
