+++ jest.config.js
@@ -9,6 +9,7 @@ module.exports = {
     '^.+\\.js$': 'babel-jest',
     '^.+\\.tsx?$': 'ts-jest',
   },
+  forceExit: true,
   testRegex: '\\.(test|spec)\\.[t]sx?$',
   moduleFileExtensions: ['ts', 'js', 'json'],
   collectCoverageFrom: ['**/integration-tests/**', '!**/generated/**'],
