+++ packages/ts-sdk-apm/test/apm.test/apm.test.ts
@@ -3,13 +3,13 @@ import {
   type MiddlewareRequestLegacy,
   type MiddlewareResponseLegacy,
   createTelemetryMiddlewareV2,
-  MiddlewareRequest,
-  OTelemetryMiddlewareOptions,
-  Next,
+  // MiddlewareRequest,
+  // OTelemetryMiddlewareOptions,
+  // Next,
 } from '../../src'
-import { recordNewrelic, time } from '../../src/helpers'
-import { createApiBuilderFromCtpClient } from '@commercetools/platform-sdk'
-import fetch from 'node-fetch'
+// import { recordNewrelic, time } from '../../src/helpers'
+// import { createApiBuilderFromCtpClient } from '@commercetools/platform-sdk'
+// import fetch from 'node-fetch'
 
 jest.mock('../../opentelemetry', () => {})
 
@@ -89,62 +89,5 @@ describe('apm', () => {
     }
 
     telemetryMiddleware(next)(request, response)
-  })
-
-  describe('custom metrics', () => {
-    const projectKey = process.env.CTP_PROJECT_KEY as string
-    const authMiddlewareOptions = {
-      host: 'https://auth.europe-west1.gcp.commercetools.com',
-      projectKey,
-      credentials: {
-        clientId: process.env.CTP_CLIENT_ID as string,
-        clientSecret: process.env.CTP_CLIENT_SECRET as string,
-      },
-      scopes: [`manage_project:${projectKey}`],
-      httpClient: fetch,
-    }
-
-    const httpMiddlewareOptions = {
-      host: 'https://api.europe-west1.gcp.commercetools.com',
-      httpClient: fetch,
-    }
-
-    it('should get request response time', async () => {
-      const _middleware = (options?: OTelemetryMiddlewareOptions) => {
-        return (next: Next) => {
-          return async (req: MiddlewareRequest) => {
-            const startTime = time()
-
-            const res = await next(req)
-
-            const endTime = time()
-            const responseTime = endTime - startTime
-            res['response_time'] = responseTime
-
-            // record stats for newrelic
-            recordNewrelic(responseTime)
-            return res
-          }
-        }
-      }
-
-      const telemetryOptions = {
-        userAgent: 'typescript-sdk-middleware-newrelic',
-        createTelemetryMiddleware: _middleware,
-      }
-
-      const client = new ClientBuilder()
-        .withTelemetryMiddleware(telemetryOptions)
-        .withAnonymousSessionFlow(authMiddlewareOptions)
-        .withHttpMiddleware(httpMiddlewareOptions)
-        .build()
-
-      const api = createApiBuilderFromCtpClient(client)
-      const result = await api.withProjectKey({ projectKey }).get().execute()
-
-      expect(typeof result).toEqual('object')
-      expect(result).toHaveProperty('response_time')
-      expect(typeof result['response_time']).toEqual('number')
-    }, 15000)
   })
 })
